Split long callback tests.
